requirements.txt
# Core dependencies
agno>=1.0.8
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.3

# Service integrations
slack-sdk>=3.23.0
pygerduty>=0.38.3
kubernetes>=29.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
